12/4 member, bus CRUD 구현 완
CRUD 구현을 1차적으로 완료 했다.
이제부터 버스 알림이의 기능
busInfo를 구현해보겠다.

+++ Mind/src/main/java/com/example/busring/dto/bus.java
@@ -7,18 +7,18 @@ public class bus {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int num;
+    private long num;
     @Column
     private String plate;
     @Column
     private String bus_name;
 
 
-    public int getNum() {
+    public long getNum() {
         return num;
     }
 
-    public void setNum(int num) {
+    public void setNum(long num) {
         this.num = num;
     }
